Threads live update added

+++ app/src/main/java/club/electro/dto/SubscriptionArea.kt
@@ -2,7 +2,7 @@ package club.electro.dto
 
 data class SubscriptionArea (
     val id: Long,
-    val type: Int,
+    val type: Byte,
     val object_id: Long,
     val name: String,
     val image: String,
